+++ app/models/forest_liana/model/collection.rb
@@ -6,13 +6,14 @@ class ForestLiana::Model::Collection
 
   attr_accessor :name, :fields, :actions, :segments, :only_for_relationships,
     :is_virtual, :is_read_only, :is_searchable, :display_name, :icon,
-    :integration, :pagination_type
+    :integration, :pagination_type, :search_fields
 
   def initialize(attributes = {})
     @actions = []
     @segments = []
     @is_searchable = true
     @is_read_only = false
+    @search_fields = nil
 
     attributes.each do |name, value|
       send("#{name}=", value)